requirements.txt
langchain
PyPDF2
python-dotenv
chainlit
langchain_google_genai 
pillow 
pydash 
pydantic 
pinecone-client 
sentence-transformers